feat: 알고리즘에 거리 계산 연결
calculateurs에서 getDistance method를 사용하도록 반영함

+++ Recommend/matchingAlgorithm/GetDistance.java
@@ -6,12 +6,13 @@ import org.neo4j.driver.Result;
 import org.neo4j.driver.Session;
 import org.neo4j.driver.Value;
 
+//neo4j 그래프 db로부터 노드간 최단거리를 가져오는 함수
 public class GetDistance {
-  public static double getDistance(String stack1, String stack2){
+  public static float getDistance(String stack1, String stack2){
     String uri = "neo4j+s://1fc435f2.databases.neo4j.io";
     String user = "neo4j";
     String password = "hgjtCyWMi5hVUIHhc0Au_VLVcOGkxBe0FM02DqUoX7k";
-    double distance=0;
+    float distance=0;
 
     try (Driver driver = GraphDatabase.driver(uri, AuthTokens.basic(user, password));
         Session session = driver.session()) {
@@ -27,7 +28,7 @@ public class GetDistance {
       Record record = result.next();
       Value totalWeight = record.get("totalWeight");
       System.out.println("Total Weight: " + totalWeight.asDouble());
-      distance=totalWeight.asDouble();
+      distance=totalWeight.asFloat();
 
     } catch (Exception e) {
       e.printStackTrace();
